+++ CHANGELOG.md
@@ -14,6 +14,8 @@
     sharing not properly configured issue #865
 * ODSettings
   * Removed AD group guid dependency issue # 862
+* TeamsTeam
+  * Added support for visibility HiddenMembership
 * MODULES
   * M365DSCUtil
     * Fixed an issue where function Test-M365DSCObjectHasProperty was missing